Add files via upload
Weather is the mix of events that happen each day in our atmosphere. Weather is different in different parts of the world and changes over minutes, hours, days and weeks. Most weather happens in the troposphere, the part of Earth’s atmosphere that is closest to the ground.
GET CURRENT WEATHER INFO USING HTML CSS JS & API